Build from loose dependencies
Building from the frozen dependencies means that we lose insight into
what our dependencies actually are. We need to be able to express what
are the actual packages needed to run Portal, and what versions are
required. This brings that in via a requirements.txt, which declares the
packages needed, and an upper-constraints.txt, which specifies version
constraints for those packages, if any. Version constraints were figured
out by looking at all the changelogs for the relevant packages and
finding versions that match our constraints (namely, Django 1.8 and
djangocms 3.1).

Also updates Dockerfile in a few places to make it save a bit of space.